modules/Empty/pyproject.toml
[tool.poetry]
name = "Automation module for Empty"
version = "0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
sekoia-automation-sdk = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
requests = "*"
requests-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "main.py",
]